+++ src/ui/index.js
@@ -3,6 +3,8 @@ export { default as Stack } from './Stack';
 export { default as Grid } from './Grid';
 export { default as Col } from './Col';
 export { default as Card } from './Card';
+export { default as Button } from './Button';
+export { default as IconButton } from './IconButton';
 export { default as Dialog } from './Dialog';
 export { DialogBackdrop, DialogPanel } from './Dialog';
 export { default as Skeleton, SkeletonBlock, SkeletonCircle } from './Skeleton';